+++ Source/Sitecore.FakeDb.NSubstitute.Tests/GettingStarted.cs
@@ -6,20 +6,6 @@
   public class GettingStarted
   {
     [Fact]
-    public void HowDoIMockAuthenticationProvider()
-    {
-      // the authentication provider is a mock created by NSubstitute;
-      // the Login() method should return 'true' when it is called with parameters 'John' and 'true'
-      Sitecore.Security.Authentication.AuthenticationManager.Provider.Login("John", true).Returns(true);
-
-      // the authentication manager is called with the expected parameters. It returns 'true'
-      Assert.True(Sitecore.Security.Authentication.AuthenticationManager.Login("John", true));
-
-      // the authentication manager is called with some unexpected parameters. It returns 'false'
-      Assert.False(Sitecore.Security.Authentication.AuthenticationManager.Login("Robber", true));
-    }
-
-    [Fact]
     public void HowDoIMockBucketProvider()
     {
       // act