+++ Source/Sitecore.FakeDb.NSubstitute/Pipelines/ReleaseFakeDb/ReleaseProviderMock.cs
@@ -3,15 +3,9 @@
   using global::NSubstitute;
   using Sitecore.Buckets.Managers;
   using Sitecore.Pipelines;
-  using Sitecore.Security.Authentication;
 
   public class ReleaseProviderMock
   {
-    public virtual void ReleaseAuthenticationProvider(PipelineArgs args)
-    {
-      AuthenticationManager.Provider.ClearReceivedCalls();
-    }
-
     public virtual void ReleaseBucketProvider(PipelineArgs args)
     {
       BucketManager.Provider.ClearReceivedCalls();