+++ Source/Sitecore.FakeDb.Tests/ContextTest.cs
@@ -1,0 +1,22 @@
+﻿namespace Sitecore.FakeDb.Tests
+{
+  using FluentAssertions;
+  using Xunit;
+
+  public class ContextTest
+  {
+    [Fact]
+    public void ShouldGetDefaultDomain()
+    {
+      // act & assert
+      Context.Domain.Name.Should().Be("default");
+    }
+
+    [Fact]
+    public void ShouldGetDefaultuser()
+    {
+      // act & assert
+      Context.User.Name.Should().Be("default\\Anonymous");
+    }
+  }
+}