+++ Source/Sitecore.FakeDb/Security/Web/FakeRoleProvider.cs
@@ -3,58 +3,60 @@ using System.Web.Security;
 
 namespace Sitecore.FakeDb.Security.Web
 {
-    public class FakeRoleProvider : RoleProvider
+  public class FakeRoleProvider : RoleProvider
+  {
+    private readonly string[] emptyRoles = { };
+
+    public override string ApplicationName { get; set; }
+
+    public override bool IsUserInRole(string username, string roleName)
     {
-        public override string ApplicationName { get; set; }
-
-        public override bool IsUserInRole(string username, string roleName)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override string[] GetRolesForUser(string username)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void CreateRole(string roleName)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override bool DeleteRole(string roleName, bool throwOnPopulatedRole)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override bool RoleExists(string roleName)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void AddUsersToRoles(string[] usernames, string[] roleNames)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override void RemoveUsersFromRoles(string[] usernames, string[] roleNames)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override string[] GetUsersInRole(string roleName)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override string[] GetAllRoles()
-        {
-            throw new NotImplementedException();
-        }
-
-        public override string[] FindUsersInRole(string roleName, string usernameToMatch)
-        {
-            throw new NotImplementedException();
-        }
+      throw new NotImplementedException();
     }
+
+    public override string[] GetRolesForUser(string username)
+    {
+      return this.emptyRoles;
+    }
+
+    public override void CreateRole(string roleName)
+    {
+      throw new NotImplementedException();
+    }
+
+    public override bool DeleteRole(string roleName, bool throwOnPopulatedRole)
+    {
+      throw new NotImplementedException();
+    }
+
+    public override bool RoleExists(string roleName)
+    {
+      throw new NotImplementedException();
+    }
+
+    public override void AddUsersToRoles(string[] usernames, string[] roleNames)
+    {
+      throw new NotImplementedException();
+    }
+
+    public override void RemoveUsersFromRoles(string[] usernames, string[] roleNames)
+    {
+      throw new NotImplementedException();
+    }
+
+    public override string[] GetUsersInRole(string roleName)
+    {
+      throw new NotImplementedException();
+    }
+
+    public override string[] GetAllRoles()
+    {
+      throw new NotImplementedException();
+    }
+
+    public override string[] FindUsersInRole(string roleName, string usernameToMatch)
+    {
+      throw new NotImplementedException();
+    }
+  }
 }